Split out Swatch context init function from generic middleware
