+++ lib/expose.js
@@ -1,6 +1,7 @@
 const Router = require('koa-router');
 const swatchjsKoa = require('swatchjs-koa-middleware');
 
+const context = require('./context');
 const defaults = require('./defaults');
 const handler = require('./handler');
 const middleware = require('./middleware');
@@ -23,7 +24,7 @@ function expose(app, methods, opts) {
       const adapterMiddleware = noAuth ? [] : [options.authAdapter];
 
       const swatchMiddleware = [
-        middleware.swatchCtx(options),
+        context.init(options),
         swatchjsKoa.logger.init,
       ];
       const adapter = swatchMiddleware.concat(adapterMiddleware);
