+++ lib/middleware.js
@@ -1,20 +1,5 @@
 const response = require('./response');
 
-function swatchCtx(options) {
-  async function swatchCtxMiddleware(koaCtx, next) {
-    // Create the swatchCtx object on the KOA ctx
-    koaCtx.swatchCtx = {};
-    koaCtx.swatchCtx.request = {
-      onException: options.onException,
-    };
-
-    // Continue chain of handlers
-    await next();
-  }
-
-  return swatchCtxMiddleware;
-}
-
 function wrapMiddleware(fn) {
   async function middlewareWrapper(koaCtx, next) {
     try {
@@ -30,6 +15,5 @@ function wrapMiddleware(fn) {
 
 
 module.exports = {
-  swatchCtx,
   wrapMiddleware,
 };
